--- src/App.js
@@ -1,148 +1,0 @@
-import { useState } from "react";
-
-// const initialItems = [
-//   { id: 1, description: "Passports", quantity: 2, packed: false },
-//   { id: 2, description: "Socks", quantity: 12, packed: true },
-//   { id: 3, description: "Charger", quantity: 1, packed: false },
-//   { id: 4, description: "Sunglasses", quantity: 1, packed: false },
-//   { id: 5, description: "Jeans", quantity: 3, packed: true },
-// ];
-
-export default function App() {
-  const [items, setItems] = useState([]);
-
-  function handleAddItems(item) {
-    setItems((items) => [...items, item]); //insert array
-  }
-
-  function handleDeleteItem(id) {
-    console.log(id);
-
-    setItems((items) => items.filter((item) => item.id !== id));
-  }
-
-  function handleToggleItem(id) {
-    setItems((items) =>
-      items.map((item) =>
-        item.id === id ? { ...item, packed: !item.packed } : item
-      )
-    );
-  }
-
-  return (
-    <div className="app">
-      <Logo />
-      <Form addItemToList={handleAddItems} />
-      <PackingList
-        itemList={items}
-        onDeleteItem={handleDeleteItem}
-        onToggleItem={handleToggleItem}
-      />
-      <Stats items={items} />
-    </div>
-  );
-}
-
-function Logo() {
-  return <h1>🌴 Far Away 🧳</h1>;
-}
-
-function Form({ addItemToList }) {
-  const [description, setDescription] = useState("");
-  const [quantity, setQuantity] = useState(1);
-
-  function handleSubmit(e) {
-    e.preventDefault(); // prevent reload
-
-    if (!description) return;
-
-    const newItem = { description, quantity, packed: false, id: Date.now() };
-    console.log(newItem);
-    addItemToList(newItem);
-
-    setDescription("");
-    setQuantity(1);
-  }
-  return (
-    <form className="add-form" onSubmit={handleSubmit}>
-      <h3>What do you need for your trip? ☺</h3>
-      <select
-        value={quantity}
-        onChange={(e) => {
-          setQuantity(Number(e.target.value));
-        }}
-      >
-        {Array.from({ length: 20 }, (_, i) => i + 1).map((num) => (
-          <option value={num} key={num}>
-            {num}
-          </option>
-        ))}
-      </select>
-      <input
-        type="text"
-        placeholder="Item..."
-        value={description}
-        onChange={(e) => {
-          setDescription(e.target.value);
-        }}
-      />
-      <button>Add</button>
-    </form>
-  );
-}
-
-function PackingList({ itemList, onDeleteItem, onToggleItem }) {
-  return (
-    <div className="list">
-      <ul>
-        {itemList.map((item) => (
-          <Item
-            item={item}
-            onDeleteItem={onDeleteItem}
-            onToggleItem={onToggleItem}
-            key={item.id}
-          />
-        ))}
-      </ul>
-    </div>
-  );
-}
-
-function Item({ item, onDeleteItem, onToggleItem }) {
-  return (
-    <li>
-      <input
-        type="checkbox"
-        value={item.packed}
-        onChange={() => onToggleItem(item.id)}
-      />
-      <span style={item.packed ? { textDecoration: "line-through" } : {}}>
-        {item.quantity} {item.description}
-      </span>
-      <button onClick={() => onDeleteItem(item.id)}>❌</button>
-    </li>
-  );
-}
-
-function Stats({ items }) {
-  if (!items.length)
-    return (
-      <p className="stats">
-        <em>Start adding some items to your packing list 🚀</em>
-      </p>
-    );
-  const numItems = items.length;
-  const numPacked = items.filter((item) => item.packed).length;
-  const percentage = Math.round((numPacked / numItems) * 100);
-
-  return (
-    <footer className="stats">
-      <em>
-        {percentage === 100
-          ? "You've got everything! Ready to go ✈"
-          : `🧳 You have ${numItems} items on your list, and you already packed
-        ${numPacked} (${percentage}%)`}
-      </em>
-    </footer>
-  );
-}
